Documentation on QuantumBridge, renaming of the file and ruff/mypy
