+++ merlin/bridge/quantum_bridge.py
